requirements.txt
phidata
yfinance
python-dotenv
duckduckgo-search
fastapi
uvicorn
groq
openai
dotenv
